Renamed Applications namespace to ApplicationModel
