+++ DotNET.Fundamentals.sln
@@ -83,11 +83,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Rules", "Specifications\Rul
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Security", "Specifications\Security\Security.csproj", "{94C75826-7DAF-4BED-A2B2-F4E3DCEEEB38}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Applications", "Source\Applications\Applications.csproj", "{15A8543B-BBA0-4B1D-8F8C-39B99245FB78}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ApplicationModel", "Source\ApplicationModel\ApplicationModel.csproj", "{15A8543B-BBA0-4B1D-8F8C-39B99245FB78}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Artifacts", "Source\Artifacts\Artifacts.csproj", "{F9A89B59-0B53-426C-A667-E458DA33332C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Applications", "Specifications\Applications\Applications.csproj", "{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Artifacts", "Specifications\Artifacts\Artifacts.csproj", "{61CE6A3A-F874-4C06-947E-18BAE5EB6743}"
 EndProject
@@ -678,18 +676,6 @@ Global
 		{F9A89B59-0B53-426C-A667-E458DA33332C}.Release|x64.Build.0 = Release|x64
 		{F9A89B59-0B53-426C-A667-E458DA33332C}.Release|x86.ActiveCfg = Release|x86
 		{F9A89B59-0B53-426C-A667-E458DA33332C}.Release|x86.Build.0 = Release|x86
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Debug|x64.ActiveCfg = Debug|x64
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Debug|x64.Build.0 = Debug|x64
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Debug|x86.ActiveCfg = Debug|x86
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Debug|x86.Build.0 = Debug|x86
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Release|x64.ActiveCfg = Release|x64
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Release|x64.Build.0 = Release|x64
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Release|x86.ActiveCfg = Release|x86
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA}.Release|x86.Build.0 = Release|x86
 		{61CE6A3A-F874-4C06-947E-18BAE5EB6743}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{61CE6A3A-F874-4C06-947E-18BAE5EB6743}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{61CE6A3A-F874-4C06-947E-18BAE5EB6743}.Debug|x64.ActiveCfg = Debug|x64
@@ -1320,7 +1306,6 @@ Global
 		{94C75826-7DAF-4BED-A2B2-F4E3DCEEEB38} = {9F86DDBC-3B12-43A9-95F6-C6AD7B6553E2}
 		{15A8543B-BBA0-4B1D-8F8C-39B99245FB78} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
 		{F9A89B59-0B53-426C-A667-E458DA33332C} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
-		{F7B202F0-0FA2-4B8C-8255-7D844F4863BA} = {9F86DDBC-3B12-43A9-95F6-C6AD7B6553E2}
 		{61CE6A3A-F874-4C06-947E-18BAE5EB6743} = {9F86DDBC-3B12-43A9-95F6-C6AD7B6553E2}
 		{F39F6574-15D4-4AFE-AB8B-67EB0A5626B5} = {9F86DDBC-3B12-43A9-95F6-C6AD7B6553E2}
 		{F94F09A6-CA08-44A1-AC96-C8D52EB3E931} = {780A06A2-8A9B-49DA-AA18-472C4F712673}
