+++ Source/ApplicationModel/Microservice.cs
@@ -4,7 +4,7 @@
 using System;
 using Dolittle.Concepts;
 
-namespace Dolittle.Applications
+namespace Dolittle.ApplicationModel
 {
     /// <summary>
     /// Represents the concept of a microservice.
