+++ Source/Execution/Execution.csproj
@@ -10,7 +10,7 @@
   </ItemGroup>
   
   <ItemGroup>
-    <ProjectReference Include="..\Applications\Applications.csproj" />
+    <ProjectReference Include="..\ApplicationModel\ApplicationModel.csproj" />
     <ProjectReference Include="..\Concepts\Concepts.csproj" />
     <ProjectReference Include="..\DependencyInversion\DependencyInversion.csproj" />
     <ProjectReference Include="..\Tenancy\Tenancy.csproj" />
