+++ Source/Execution/ExecutionContext.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Globalization;
-using Dolittle.Applications;
+using Dolittle.ApplicationModel;
 using Dolittle.Concepts;
 using Dolittle.Security;
 using Dolittle.Tenancy;
@@ -18,7 +18,7 @@ namespace Dolittle.Execution
         /// Initializes a new instance of the <see cref="ExecutionContext"/> class.
         /// </summary>
         /// <param name="application"><see cref="Application"/> that is currently executing.</param>
-        /// <param name="microservice"><see cref="Applications.Microservice"/> that is currently executing.</param>
+        /// <param name="microservice"><see cref="ApplicationModel.Microservice"/> that is currently executing.</param>
         /// <param name="tenant"><see cref="TenantId"/> that is currently part of the <see cref="ExecutionContext"/>.</param>
         /// <param name="environment"><see cref="Environment"/> for this <see cref="ExecutionContext"/>.</param>
         /// <param name="correlationId"><see cref="CorrelationId"/> for this <see cref="ExecutionContext"/>.</param>
